requirements.txt
SQLAlchemy~=2.0.39
fastapi~=0.115.11
aiofiles~=24.1.0
python-jose~=3.4.0
passlib~=1.7.4
pydantic~=2.10.6
uvicorn~=0.34.0
python-dotenv~=1.1.0
langchain-community~=0.3.20
langchain-text-splitters~=0.3.7
sentence-transformers~=4.0.1
chromadb~=0.6.3
langchain~=0.3.22
pydantic-settings~=2.8.1
psycopg2-binary==2.9.9
